Document exposition format directly on the website.
This only documents the latest 0.0.4 exposition format version, with a
link to the legacy doc at the bottom. I also took the liberty to
fix up/adapt some of the content. Tables from the original doc are
transliterated as nested lists instead, because Markdown doesn't support
lists in tables (and the old doc made heavy use of them). A future
option could be to go for completely manual HTML tables with HTML lists
in them.

This fixes https://github.com/prometheus/docs/issues/28