Add and fix a number of minor things.
- Sort exporters lexicographically (but still put the node exporter first).
- Call "unofficial" exporters "independently maintained" exporters and
  adjust the wording around it.
- Add the resurrected Munin exporter.
- Add the crucial information about RAM to the benchmark results.
- Add a new "media" section for introductory blog posts and talks.
- Refer to that section instead of the SC blog post in the Overview.

Note about Stapelberg's talk: As it is in German and without
subtitles, it is only approachable for a limited audience, but it is
in fact a very good introduction, so I opted to list it here in last
position. Measured in "number of Wikipedia articles", German is the
2nd most important language on the Internet after English. So I hope
it's not completely obscure.