Better explain file_sd's expected format
Indicate in the docs that the files need to contain an array of target groups, not a standalone one. This was only discoverable by looking at the output of prometheus saying the file was bad.

I know the docs indicate "zero or more" which implies it is an array, but that detail is extremely easy to miss.