Add download section
This adds a new download section to the page with list for all available
precompiled versions for officially maintained components. The release
information are fetched via the Github API and rendered as static list.

Currently only the latest release and the newest pre-release (if
available) are show. This could be extended later. In order to keep the
list usable besides the massive list of prebuilt binaries, a javascript
os/arch filter has been added, defaulting to popular versions.

The following changes have been made on the way:

* Ensure all pages have a footer.
* Remove duplicated jquery inclusion.
* Use CSS to define text case of jumbotron buttons.
* Add Download section to the header.
* Reduce Github link to just the logo to save space.
* Update Getting Started and Installing sections.