Use authentification for github release API requests
So far the requests against the Github API to fetch release information
has been made unauthenticated, making them subject to a rate-limit of 10
requests per hour per IP. As the build process is executed on Travis CI,
this results in many build failures.

The requests are now made with a personal access token, in the form of
GITHUB_AUTHENTICATION="-u prombot:<oauth token>". The standard printing
in the Makefile has been removed in order to not expose the token.