[WIP] introduce categories to exporterlist
As @juliusv mentioned once in #prometheus, the list of external exporters has grown quite a bit. This attempts to increase overview by introducing categories for the third party exporters.

The chosen categories are a first proposal and might still be subject to change, depending on the feedback, any of which is greatly appreciated!