Point to default port list. (#692)
It has unintentionally become a defacto listing of exporters, some of whom
we may not wish to list here for various reasons or for which noone has
sent a PR yet. This helps make it easier for users to find exporters,
and potentially aid in development of the ones that are still maturing.