Add a PHP-FPM exporter to the list of exporters
Currently, there is a large set of exporters available for a lot of different software tools.
However, the popular runtime "PHP-FPM", used for executing PHP script through the
FastCGI protocol, is not listed.

Various exporters currently exist, and the author has not tested any (including the
suggested) extensively. However, a quick review of the available options shows that,
in the authors limited review, this appears to be well constructed, thought out and
documented.

This commit adds the exporter to the canonical list for easy discovery. 

Design Notes:

Alternatives considered were:
- https://github.com/peakgames/php-fpm-prometheus/ (no releases, less docs)
- https://github.com/blablacar/phpfpm-prometheus-exporter (no docs)