Shift the PHP-FPM exporter to the "misc" section
Previously, the author had added a new section titled "interpreters", and
added the PHP-FPM exporter to this section under the belief that, given there
was none (that is, neither node or python are represented) this could be a
separate class of application. However, with review it looks like the better
place for this exporter is under the "misc" section.

This commit moves the PHP-FPM exporter to this section.