Add a PHP-FPM exporter to the list of exporters (#791)
Currently, there is a large set of exporters available for a lot of different software tools.
However, the popular runtime "PHP-FPM", used for executing PHP script through the
FastCGI protocol, is not listed.

Various exporters currently exist, and the author has not tested any (including the
suggested) extensively. However, a quick review of the available options shows that,
in the authors limited review, this appears to be well constructed, thought out and
documented.

This commit adds the exporter to the canonical list for easy discovery. 

Design Notes:

Alternatives considered were:
- https://github.com/peakgames/php-fpm-prometheus/ (no releases, less docs)
- https://github.com/blablacar/phpfpm-prometheus-exporter (no docs)

* Shift the PHP-FPM exporter to the "misc" section

Previously, the author had added a new section titled "interpreters", and
added the PHP-FPM exporter to this section under the belief that, given there
was none (that is, neither node or python are represented) this could be a
separate class of application. However, with review it looks like the better
place for this exporter is under the "misc" section.

This commit moves the PHP-FPM exporter to this section.