Suffix the groupKey with a `,` character
Previously, the author amended the groupKey type from an int
to a string. In reviewing this change, the author noted that, although
the string is not syntactically correct json, all other type descriptors are
suffixed with a "," character.

This commit adds the "," character simply for consistency, and because
once the author noticed it they felt compelled to fix it to silence the
code inconsistency demons in their head.