Add note about long term metric storage
I feel the current comparison is **highly** misleading. After reading this page an argument is made that Prometheus is superior and most solutions out there are inferior. 

Long-term metrics, by-design, are not handled by prometheus. You can capture a couple of weeks of data, but not years, there is no downsampling options and memory usage is quite high. 

I feel is is EXTREMELY important to highlight this design decision here so sysadmins of the world do not get confused.