Okay - third time lucky.
It's still not clear to me why the Concepts page differs from the
Glossary but maybe I am missing something?

If I am - perhaps we could jump onto a quick chat. I feel like these
definitions of the abstractions is super important and I am not sure
they are clear.