Fix link to instrumentation practices docs
It seems that the documentation has been rearranged at some time in the past and not all links to other documentation pages point to their intended locations.  This change fixes the link to the instrumentation practices documentation.  This pull request is submitted in the hope that it is helpful.  If you want anything changed I will gladly update it and resubmit.