Replace lovoo/ipmi_exporter by soundcloud/ipmi_exporter (#1029)
Frankly, I would prefer to list both as they have quite different
approaches. Lovoo's exporter runs on the IPMI-supervised host and uses
the `dev/ipmi0` device to server metrics. This is straight-forward but
doesn't work if the host is powered down. However, in some secnarios,
it's the whole point of IPMI to be able to work with machines while
they are powered down or stuck in the boot process. If you need
metrics in those scenarios, too, the SoundCloud exporter works
better. It uses the blackbox exporter approach and is thus (arguably)
a tad more complicated to set up.

If I _had_ to choose one and only one IPMI exporter, I would pick the
SoundCloud one as it covers all scenarios.  Since @brian-brazil
insisted so far to only ever list one solution for a certain thing to
export, this commit _replaces_ the Lovoo exporter by the SoundCloud
exporter.

Discuss!

Signed-off-by: beorn7 <beorn@soundcloud.com>