Added guide about proxy exporters
This guide should answer a lot of common questions on the mailing list
and help people understand how this type of exporters work while also
giving them practical experience.

It handles a few common pitfalls and is written to be as accessible as
possible. Hopefully it will help many people to have a less frustrating
experience using Prometheus and free the mailing list from some load of
often repeated questions.

It is a prerequirement for an SNMP exporter guide, that is useful for beginners.

Signed-off-by: Nicolai von Neudeck <nicolai.von.neudeck@sap.com>
Signed-off-by: Nicolai von Neudeck <nicolai@vonneudeck.com>