Adding Thanos to integration of remote write. (#1358)
Thanos mainly uses remote read to integrate with Prometheus. Recently Thanos receiver was added so we can put it as `write` integration as well. I added `read as client` as well as to me this is important detail. Thanos integrates with Prometheus as remote read client, so technically it interates with read? Thanos does not expose remote read on it's own though.

Thoughts? (:

Signed-off-by: Bartek Plotka <bartek@improbable.io>