Spring-cleaning
It has been a while hence, a few modifications became necessary.

Changed wording to be less emotional.
Changed `"` to `'` where recommended by RichiH.
Added `sudo` to commands to be less frustrating. 
Changed guide to reflect changes in Dockers’ behaviour under Linux. 
Changed link to an actual query at the end to be more entertaining.
Changed person to `you` to clarify something the reader is expected to do.
Tried all commands and tested if they still work as intended.

Signed-off-by: Nicolai von Neudeck <nicolai@vonneudeck.com>