Improve robustness of git working tree checkout (#1764)
For unknown reasons the netlify pipeline ends up in a state where a git
worktree is not created after a fresh clone of the repository. The
previous implementation checked for the actual checked out working tree
and assumed the administrative git worktree files have to exist as well
in that case. This change explicitly verifies the existence of both path
structures and handles more cases of inconsistencies gracefully.

Signed-off-by: Tobias Schmidt <tobidt@gmail.com>