Remove aaS entry from software list
This is a list of software projects and products implemented our standard, not offers based on said software.

Initial internal discussion on if we want to carry a list of offerings has started, but not gotten anywhere yet.

Brian remembers digging to see if it was based on Cortex and merging because it was not obviously based on software listed already. As such, this commits reverts a mistake.

Signed-off-by: Richard Hartmann <richih@richih.org>