Add GSuite to onboarding
_This is an informational change and therefore doesn't go through the
formal procedure of a governance change._

This also mentions the team's calendar. Note that those calendars are
notoriously difficult to find, so whoever onboards the new team member
should give them a direct link to it. (But I didn't want to get more
technical than I did in this PR.)

Signed-off-by: beorn7 <beorn@grafana.com>