instrumenting/clientlibs: add process_threads
This reserves a new well-known `process_threads` gauge for client
libraries, in order to expose OS threads count on all instrumented
processes. Similar metrics are already exposed by Go and JVM
libraries, and this has been recently requested for Rust too.
As this metrics is runtime independent, it makes sense to have it
generally uniform across languages and libraries.

Signed-off-by: Luca BRUNO <luca.bruno@coreos.com>