instrumenting/clientlibs: add process_threads
This reserves a new well-known `process_threads` gauge for client
libraries, in order to expose OS threads count on all instrumented
processes.
As this metrics is language/runtime independent, it makes sense to
be under the `process_` namespace so that libraries can align on
the name (if/when they start exposing this).

Signed-off-by: Luca BRUNO <luca.bruno@coreos.com>