Add link to docs about `up` metric
I read this section a few times, but was still confused about how I'd alert or graph uptime if the app or exporter doesn't export a 1/0 metric. The missing piece of information is that there's a built-in `up` metric. This PR adds a link so the next guy or gal doesn't get confused like I did. :-)

Signed-off-by: Michael Kopinsky <Michael.Kopinsky@pennmedicine.upenn.edu>