Solve contradictive metrics naming recomendation
For metrics specific to an application it's recommended to put the application name as a prefix and for all metrics, it's recommended not to put labels into metrics names. After the discussion in the [issue](https://github.com/prometheus/docs/issues/2167) it appeared that for metrics specific to applications it's ok.

Signed-off-by: Aleksey Bobyr <agat00@gmail.com>