Add various references to native histograms
This is not yet an exhaustive documentation. The changes here are
mostly meant to avoid confusion.

Once native histograms are closer to becoming a stable feature, we
will be able (and should) provide more detailed documentation.

Note that additional documentation is provided as part of the v2.40
release in the prometheus/prometheus repository.

Signed-off-by: beorn7 <beorn@grafana.com>