Rename "Prometheus" docs section to "Prometheus Server" (#2473)
As per Björn's comment on the team mailing list:

"I think the current "Prometheus" section is not very well named. It is
essentially the part coming from the prometheus/prometheus repo (not
prometheus/docs, as the rest fo the site), and it means "The
Prometheus Server". I think it should be named that way because just
"Prometheus" could mean a lot of things. On prometheus.io,
"Prometheus" mostly refers to the overarching project."

This also adds CSS styles to make sure that the "Prometheus Server" text gets
line-broken in such a way on narrow displays that the second line is indented
correctly.

Signed-off-by: Julius Volz <julius.volz@gmail.com>