Rename "Alerting" to "Alertmanager"
After #2473 finally reaming the "Prometheus" section to "Prometheus
Server", I think we should also rename the top-level "Alerting"
section to "Alertmanager". This section is just the embedded
documentation from the Alertmanager repository, and it speaks volumes
that the "Precise alerting" link on the landing page leads to
https://prometheus.io/docs/prometheus/latest/configuration/alerting_rules/
rather than the so-called "Alerting" section.

In defense of the current section name, there is an "Alerting
Overview" subsection that links to other places, but I think it's
still better to call the whole section "Alertmanager', also
because it makes it easier to find the actual Alertmanager
documentation if you are looking specifically for it (rather than
Alerting in general).

Final point: The first item if you expand the "Alerting" navbar is the
selection of the Alertmanager version, which is also confusing if the
title just above it just reads "Alerting".

Signed-off-by: beorn7 <beorn@grafana.com>