Add native histogram specification (#2539)
Add native histogram specification

It is partially more documentation than specification, but I assume
the specification section is still the least bad place for it.

This is intentionally not yet complete. TODOs mark the missing parts,
most of which cannot be resolved at the moment because implementation
work is still missing and/or decisions still have to be made.

The intention here is to finally get something out here, even if it is
pretty much still a "living document". It will probably take us
another year or so to resolve all TODOs.

Signed-off-by: beorn7 <beorn@grafana.com>

---------

Signed-off-by: beorn7 <beorn@grafana.com>