Add a "tl;dr:" style note at the top of the security page (#2558)
The security model is written in a very technical form that is hard to
understand for naive users. Naive users are, sadly, also most prone to
run Prometheus components exposed to the internet. I have little hope
that those users will read this page at all, but if they do, they
should get the message. Furthermore, we have had multiple incidents
where 3rd parties "uncover" "vulnerabilities and security flaws".
There is nothing to uncover, and the flagged behaviors are by design.
We have to make that more clear.

This partially addresses #2201, but I think we need to do more. In
particular, we should add a section about /pprof (and why we have it
on by default), be clearer about DoS'ing ("It is more likely that a
component will be accidentally taken out by a trusted user than by
malicious action." can easily be understood in an unintended way), and
provide more context about the usage of the word "untrusted".

Signed-off-by: beorn7 <beorn@grafana.com>