Add a contributor's guide
This replaces the section on the community page.

The information from the community page is not only moved into the new
contributor's guide, but also updated to the current realities. Plus,
a lot of new information is added, like a more helpful introduction
for new contributors, a coding style guide, guidelines about handling
PRs, and more. I tried to write down what I percieve as the tribal
knowledge. However, people might have different versions of this
tribal knowledge in their heads, so please discuss during the review
what we want to codify here.

Signed-off-by: beorn7 <beorn@grafana.com>