Add ICMP Exporter
ICMP Exporter is a very simple exporter to monitor latency to target hosts. It is designed to be flexible and implement features other exporters currently don't, the main one being (and the entire reason it was created in the first place) being able to specify the source interface for a ping. It can also specify the TTL of pings, useful when monitoring links between routers and alike.

Signed-off-by: Mattie Nickson <mnickson@sidingsmedia.com>