Fix flickering dropdown, clean up popover state handling
The main change here is the detaching / hiding of the dropdown when the
target element (the burger icon) is hidden from the screen. This can happen
when opening the dropdown at a screen width where the burger is shown, then
changing the screen width (e.g. by rotating your phone) to a width where
the burger is hidden, and then rotating back to where it's shown. Then the
popup would flicker around wildly after that, and a page reload would be
required to use the burger normally again. With this change, the dropdown
simply stays open (until dismissed) even if the burger disappears.

Signed-off-by: Julius Volz <julius.volz@gmail.com>