fix(om2): histograms and negative observed values (#2627)
OM1.0 required that the Sum of Histograms is not represented when there are negative observations in a histogram.

This PR is removing this requirement in OM2.0. Due to:
The requirement was never implemented by the Go and Java instrumentation libraries. Enforcing it now would be breaking.
The requirement makes it impossible to implement the use case where the user wants to measure the Sum anyway. Which means for example that you'll not be able to calculate average from Sum/Count.
We already warned users in the documentation about the possibility of Sum decreasing and not being usable for rate() 10 years ago: PR.
And native histograms will not take Sum into account when calculating counter resets during rate() , thus this problem won't come up.

Note1: the python reference implementation did follow the requirement.

Note 2: this PR does not make Sum mandatory, that is a different question.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>