Update 2025-10-30-non-code-contribution.md
I noticed that the hyperlink to the blog content guide was leading to a 404 error on the live site, even though it worked correctly on GitHub. I think the issue was caused by using a relative link. I’ve updated the link to use an absolute URL so it resolves properly in the published blog.

Signed-off-by: Victoria Nduka <122698422+nwanduka@users.noreply.github.com>