Fix scrolling to fragment identifiers on newly loaded large pages (#2788)
When following a link to a fragment (#foo) on a long documentation page,
the browser would try to scroll to the fragment too early, before some
additional client-side layout shifts (could be due to font loading,
highlighting, etc.) would cause the scroll position to be wrong again. This
component waits for the layout to settle and then manually scrolls to the
requested fragment again.

You can see the problem by clicking on:

https://prometheus.io/docs/prometheus/latest/configuration/configuration/#relabel_config

It usually does not scroll directly to the relabeling section, but
somewhere above it.

Signed-off-by: Julius Volz <julius.volz@gmail.com>